fix: silence some warnings from some versions of ICX
There is a -Wpre-c11-compat warning in 2024.2 (latest version at time
of writing) which is emitted when icx sees a _Noreturn keyword, or
calls _Static_assert.  It's not a default, and really only a problem
because I'm trying to support compiling with -Weverything without any
warnings.

This just tweaks a couple of macros (HEDLEY_NO_RETURN and
HEDLEY_STATIC_ASSERT) to wrap the output in pragmas to temporarily
disable the warning.